requirements.txt
streamlit
requests
pydantic
python-dotenv
gspread
google-auth
